added new variables and for loop for data map values
